refactor: split components into folders
