--- src/Header.css
@@ -1,15 +1,0 @@
-.container {
-    display: flex;
-    justify-content: space-between;
-    padding: 5px 10px;
-    align-items: center;
-}
-
-.navList {
-    display: flex;
-    list-style: none;
-}
-
-.navItem {
-    padding: 5px;
-}
