+++ src/Header/Header.js
@@ -1,0 +1,47 @@
+import React from "react";
+import "./Header.css";
+import {Logo} from "loft-taxi-mui-theme";
+import {Button, Container, Paper} from "@material-ui/core";
+import PropTypes from 'prop-types';
+
+const Header = ({customNavigation}) => {
+    const buttonList = [
+        {
+            name: "map",
+            text: "Карта",
+        },
+        {
+            name: "profile",
+            text: "Профиль",
+        },
+        {
+            name: "login",
+            text: "Войти",
+        },
+    ];
+    return (
+        <Paper elevation={4} tag="header" >
+            <Container>
+                <div className="container">
+                    <div className="logo">
+                        <Logo/>
+                    </div>
+                    <nav className="nav">
+                        <ul className="navList">
+                            {buttonList.map(({name, text}) => (
+                                <li className="navItem" key={name}>
+                                    <Button onClick={() => customNavigation(name)}>{text}</Button>
+                                </li>
+                            ))}
+                        </ul>
+                    </nav>
+                </div>
+            </Container>
+        </Paper>
+    );
+}
+Header.prototype = {
+    customNavigation: PropTypes.func
+}
+
+export default Header;
