--- src/Login.css
@@ -1,17 +1,0 @@
-.login {
-    background-image: url("image/login-background.jpg");
-    background-size: cover;
-    height: 100vh;
-}
-
-.loginForm {
-    padding: 44px 60px;
-    min-width: 500px;
-    margin-top: 48px;
-    margin-bottom: 48px;
-    color: rgba(0, 0, 0, 0.87);
-    transition: box-shadow 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-    background-color: #fff;
-    border-radius: 4px;
-}
-
